requirements.txt
requests
beautifulsoup4
pycryptodome
lxml